fix mock interface, reorg dir
